+++ packages/core/src/components/MenuItem.tsx
@@ -19,17 +19,19 @@ interface MenuItemProps {
 }
 
 const MenuItem: React.FC<MenuItemProps> = ({ checked = false, children, icon = null, isDisabled = false, onClick }) => (
-    <li
+    <button
         className={classNames({
             'rpv-core__menu-item': true,
             'rpv-core__menu-item--disabled': isDisabled,
         })}
+        role="menuitem"
+        type="button"
         onClick={onClick}
     >
         <div className="rpv-core__menu-item-icon">{icon}</div>
         <div className="rpv-core__menu-item-label">{children}</div>
         <div className="rpv-core__menu-item-check">{checked && <CheckIcon />}</div>
-    </li>
+    </button>
 );
 
 export default MenuItem;